Proof-of-concept for clean LuaJIT build
I can confirm that, with this patch, lpeglabel builds with no
errors or warnings under LuaJIT.

This also restores the static library target, I don't know whether
that was removed from lpeg or within lpeglabel but it happens to
be the target I need.

This isn't careful, or intended for merge, it merely provides the
missing macro.